[DP-23] - Вынес класс OnboardingPresenter в отдельный файл. Удалил протокол OnboardingViewProtocol
