+++ DavaiPosmotrim/OnboardingFlow/OnboardingPresenterProtocol.swift
@@ -10,19 +10,3 @@ import Foundation
 protocol OnboardingPresenterProtocol: AnyObject {
     func onboardingFinish()
 }
-
-final class OnboardingPresenter: OnboardingPresenterProtocol {
-
-    // MARK: - Public Properties
-    weak var coordinator: OnboardingCoordinator?
-    weak var view: OnboardingViewProtocol?
-
-    init(coordinator: OnboardingCoordinator) {
-        self.coordinator = coordinator
-    }
-
-    func onboardingFinish() {
-        guard let coordinator else { return }
-        coordinator.finish()
-    }
-}
