+++ DavaiPosmotrim/OnboardingFlow/OnboardingSceneFactory.swift
@@ -11,7 +11,6 @@ struct OnboardingSceneFactory {
     static func makeOnboardingViewController(with coordinator: OnboardingCoordinator) -> OnboardingViewController {
         let presenter = OnboardingPresenter(coordinator: coordinator)
         let viewController = OnboardingViewController(presenter: presenter)
-        presenter.view = viewController
         return viewController
     }
 }
