--- DavaiPosmotrim/OnboardingFlow/OnboardingViewProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  OnboardingViewProtocol.swift
-//  DavaiPosmotrim
-//
-//  Created by Aleksandr Garipov on 26.04.2024.
-//
-
-import Foundation
-
-protocol OnboardingViewProtocol: AnyObject {
-    // TODO: - add code to use viewControllers method in presenter
-}
